requirements.txt
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
decorator==5.0.9
google-api-core==1.31.0
google-api-python-client==2.12.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
importlib-metadata==4.6.1
isodate==0.6.0
joblib==1.0.1
nltk==3.6.2
nnvm==0.8.0
numpy==1.21.0
oauthlib==3.1.1
packaging==21.0
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.4
pyparsing==2.4.7
python-dotenv==0.18.0
pytz==2021.1
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
retry==0.9.2
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
threadpoolctl==2.2.0
topi==0.6.dev0
tqdm==4.61.2
tvm==0.6.dev0
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
XlsxWriter==1.4.4
zipp==3.5.0
